Build(code-coverage): Migrate `coverageThreshold` to `BigDecimal`
This commit updates the `CodeCoveragePlugin` to use `BigDecimal` for the code coverage threshold instead of `Int`. This allows for more precise threshold values and avoids manual type conversions during the JaCoCo violation rule configuration.

Key changes:
- **Plugin Logic**: Updated `CodeCoveragePlugin` to divide the threshold by 100 directly as a `BigDecimal`.
- **Configuration**: Modified `build.gradle.kts` to map the `CODE_COVERAGE_THRESHOLD` environment variable to `BigDecimal`, defaulting to `80.toBigDecimal()`.
- **Extension**: Changed the `coverageThreshold` property type in `CodeCoverageExtension` from `Property<Int>` to `Property<BigDecimal>`.
- **CI/CD**: Simplified step names in the GitHub Action `code-coverage-action` from "Archive coverage report" to "Archive report" and "Publish code coverage report" to "Publish report".

+++ build-logic/code-coverage/src/main/kotlin/md/vnastasi/plugin/codecoverage/CodeCoverageExtension.kt
@@ -7,6 +7,7 @@ import org.gradle.api.provider.ListProperty
 import org.gradle.api.provider.Property
 import org.gradle.kotlin.dsl.listProperty
 import org.gradle.kotlin.dsl.property
+import java.math.BigDecimal
 import javax.inject.Inject
 
 abstract class CodeCoverageExtension @Inject constructor(objectFactory: ObjectFactory) {
@@ -15,7 +16,7 @@ abstract class CodeCoverageExtension @Inject constructor(objectFactory: ObjectFa
 
     val reportDirectory: DirectoryProperty = objectFactory.directoryProperty()
 
-    val coverageThreshold: Property<Int> = objectFactory.property()
+    val coverageThreshold: Property<BigDecimal> = objectFactory.property()
 
     val excludedModules: ListProperty<Project> = objectFactory.listProperty()
